third_party/blink/web_tests/platform/generic/fast/dom/MutationObserver/shadow-dom-expected.txt

Test that MutationObservers operate in Shadow DOM

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Mutations in shadow DOM should have been observed:
PASS mutations.length is 4
PASS mutations[0].type is "attributes"
PASS mutations[1].type is "childList"
PASS mutations[2].type is "characterData"
PASS mutations[3].type is "childList"

Observing from outside shadow DOM should not see mutations in the shadow:
PASS mutations.length is 0
PASS successfullyParsed is true

TEST COMPLETE
